requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
xmltodict==0.12.0
